Added PriceFeed.sol contract
The PriceFeed.sol contract abstracts away the oraclize call.
The separation from the main contracts allows for better mocking when it comes
to unit testing. I am not quite sure it is the best approach as it also involves additional
gas costs. I think implementation will let me decide whether or not this is the correct approach we want
and whether or not we want to go for another pattern.